# swagger.yaml
...
openapi: 3.0.0
info:
  title: API Documentation
  version: 1.0.0
  description: API documentation for Restaurant management
servers:
  - url: http://localhost:3000
    description: Development server
paths:
...
  /admin/dish/create:
    post:
      summary: Thêm món ăn mới
      description: Thêm một món ăn mới vào hệ thống quản lý nhà hàng.
      tags:
        - Dish
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: "Cocacola"
                  description: Tên của món ăn
                description:
                  type: string
                  example: "Nước ngọt"
                  description: Mô tả của món ăn
                price:
                  type: number
                  example: 3
                  description: Giá của món ăn
                category:
                  type: string
                  example: "Drink"
                  description: Danh mục của món ăn
                file:
                  type: string
                  format: binary
                  description: File ảnh của món ăn
      responses:
        '201':
          description: Món ăn được thêm thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Sản phẩm Cocacola được thêm thành công"
                  data:
                    type: object
                    properties:
                      name:
                        type: string
                        example: "Cocacola"
                      description:
                        type: string
                        example: "Nước ngọt"
                      price:
                        type: number
                        example: 3
                      category:
                        type: string
                        example: "Drink"
                      imageUrl:
                        type: string
                        example: "https://res.cloudinary.com/yourcloud/image/upload/v1731144323/vmk2t4iz1pdqqrh55mf9.jpg"
                      like:
                        type: number
                        example: 0
                      deleted:
                        type: boolean
                        example: false
                      _id:
                        type: string
                        example: "672f2a84599a4f9a3dc62d54"
        '400':
          description: Yêu cầu không hợp lệ, ví dụ thiếu file hoặc món ăn đã tồn tại
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Món ăn đã tồn tại"
        '500':
          description: Thêm món ăn thất bại do lỗi server
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Thêm món ăn thất bại!"
                  error:
                    type: string
                    example: "Internal server error message"
  /admin/dish/edit/{slugDish}:
    patch:
      summary: Cập nhật thông tin món ăn
      description: Cho phép admin cập nhật thông tin của một món ăn dựa trên `slugDish`.
      tags:
        - Dish
      parameters:
        - name: slugDish
          in: path
          required: true
          description: "Slug của món ăn cần cập nhật"
          schema:
            type: string
            example: "ga-ran"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                description:
                  type: string
                  example: "Món gà rán giòn tan với vị mới"
                  description: "Mô tả mới của món ăn"
                price:
                  type: number
                  example: 10
                  description: "Giá mới của món ăn"
                category:
                  type: string
                  example: "Main Course"
                  description: "Loại món ăn mới"
                  enum: ["Appetizer", "Main Course", "Dessert", "Drink"]
                imageUrl:
                  type: string
                  example: "https://link-to-updated-image.com/image.jpg"
                  description: "URL hình ảnh mới của món ăn"
      responses:
        200:
          description: Cập nhật món ăn thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Cập nhật món ăn thành công!"
                  dish:
                    type: object
                    properties:
                      description:
                        type: string
                        example: "Món gà rán giòn tan với vị mới"
                      price:
                        type: number
                        example: 55000
                      category:
                        type: string
                        example: "Main Course"
                      imageUrl:
                        type: string
                        example: "https://link-to-updated-image.com/image.jpg"
        404:
          description: Món ăn không tồn tại
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Món ăn không tồn tại!"
        501:
          description: Lỗi trong quá trình cập nhật
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Error!"
                  error:
                    type: string
                    example: "Mô tả lỗi"
  /admin/dish/list:
    get:
      summary: Lấy danh sách các món ăn
      description: Trả về danh sách tất cả các món ăn chưa bị xóa trong hệ thống.
      tags:
        - Dish
      responses:
        200:
          description: Thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: "Success!"
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        name:
                          type: string
                          example: "Gà rán"
                          description: "Tên món ăn"
                        slug:
                          type: string
                          example: "ga-ran"
                          description: "Slug của món ăn"
                        description:
                          type: string
                          example: "Món gà rán giòn tan"
                          description: "Mô tả món ăn"
                        price:
                          type: number
                          example: 50000
                          description: "Giá của món ăn"
                        category:
                          type: string
                          example: "Main Course"
                          description: "Loại món ăn"
                          enum: ["Appetizer", "Main Course", "Dessert", "Drink"]
                        imageUrl:
                          type: string
                          example: "https://link-to-image.com/image.jpg"
                          description: "URL hình ảnh món ăn"
                        like:
                          type: integer
                          example: 12
                          description: "Số lượng lượt thích"
                        deleted:
                          type: boolean
                          example: false
                          description: "Trạng thái đã xóa hay chưa"
        400:
          description: Lỗi trong quá trình lấy danh sách món ăn
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 400
                  message:
                    type: string
                    example: "Error!"
                  error:
                    type: string
                    example: "Mô tả lỗi"
  /admin/dish/detail/{slugDish}:
    get:
      summary: Lấy chi tiết món ăn theo slug
      description: Trả về chi tiết món ăn dựa trên slug của món ăn, bao gồm tên, mô tả, số lượt thích, giá, hình ảnh và loại món ăn.
      tags:
        - Dish
      parameters:
        - name: slugDish
          in: path
          description: Slug của món ăn
          required: true
          schema:
            type: string
      responses:
        200:
          description: Thành công, trả về chi tiết món ăn
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Success!"
                  data:
                    type: object
                    properties:
                      name:
                        type: string
                        example: "Gà rán"
                        description: "Tên món ăn"
                      slug:
                        type: string
                        example: "ga-ran"
                        description: "Slug của món ăn"
                      description:
                        type: string
                        example: "Món gà rán giòn tan"
                        description: "Mô tả món ăn"
                      price:
                        type: number
                        example: 50000
                        description: "Giá của món ăn"
                      category:
                        type: string
                        example: "Main Course"
                        description: "Loại món ăn"
                        enum: ["Appetizer", "Main Course", "Dessert", "Drink"]
                      imageUrl:
                        type: string
                        example: "https://link-to-image.com/image.jpg"
                        description: "URL hình ảnh món ăn"
                      like:
                        type: integer
                        example: 12
                        description: "Số lượng lượt thích"
        400:
          description: Lỗi khi không tìm thấy món ăn
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Error!"
                  error:
                    type: string
                    example: "Không tìm thấy món ăn với slug này."
  /admin/dish/delete/{slugDish}:
    delete:
      summary: Xóa món ăn theo slug
      description: Đánh dấu món ăn là đã xóa bằng cách thiết lập trường `deleted` thành `true` trong cơ sở dữ liệu.
      tags:
        - Dish
      parameters:
        - name: slugDish
          in: path
          description: Slug của món ăn cần xóa
          required: true
          schema:
            type: string
      responses:
        200:
          description: Thành công, trả về món ăn đã bị xóa
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Xóa món ăn thành công!"
                  dish:
                    type: object
                    properties:
                      name:
                        type: string
                        example: "Gà rán"
                        description: "Tên món ăn"
                      slug:
                        type: string
                        example: "ga-ran"
                        description: "Slug của món ăn"
                      description:
                        type: string
                        example: "Món gà rán giòn tan"
                        description: "Mô tả món ăn"
                      price:
                        type: number
                        example: 50000
                        description: "Giá của món ăn"
                      category:
                        type: string
                        example: "Main Course"
                        description: "Loại món ăn"
                        enum: ["Appetizer", "Main Course", "Dessert", "Drink"]
                      imageUrl:
                        type: string
                        example: "https://link-to-image.com/image.jpg"
                        description: "URL hình ảnh món ăn"
                      like:
                        type: integer
                        example: 12
                        description: "Số lượng lượt thích"
                      deleted:
                        type: boolean
                        example: true
                        description: "Trạng thái món ăn (đã bị xóa)"
        400:
          description: Lỗi khi không tìm thấy món ăn hoặc không thể xóa
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Error!"
                  error:
                    type: string
                    example: "Không tìm thấy món ăn để xóa."
...